cleaned filed structure
