--- src/components/CourseInfoComponent
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const Card - () => {
-    return (
-        <div className="card">
-        <div className="header"> Header </div>
-        <div className="content"> content</div>
-        </div>
-    )
-}
-
-export default Card;
