+++ src/screens/App.tsx
@@ -1,42 +1,21 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
 import LandingPage from "../screens/LandingPage/LandingPage";
-import CourseSearch from "../screens/CourseSearch";
-import SavedCourses from "../screens/SavedCourses";
-import CourseLoadTester from "../screens/CourseLoadTester";
+import CourseSearch from "./CourseSearch/CourseSearch";
+import CourseLoadTester from "./CourseLoadTester/CourseLoadTester";
 import CoffeeChat from "../screens/CoffeeChat/CoffeeChat";
 import ExploreCoffeeChat from "../screens/CoffeeChat/ExploreCoffeeChat";
 import RegisterCoffeeChat from "../screens/CoffeeChat/RegisterCoffeeChat";
-import RegisterPlaceholder from "../screens/CoffeeChat/RegisterPlaceholder";
 import StudentContact from "../screens/CoffeeChat/StudentContact";
 
 const App: React.FC = () => {
   return (
     <Router>
       <div>
-        {/* <nav>
-          <ul>
-            <li>
-              <Link to="/">Landing Page</Link>
-            </li>
-            <li>
-              <Link to="/courseSearch">Course Search</Link>
-            </li>
-            <li>
-              <Link to="/savedCourses">Saved Courses</Link>
-            </li>
-            <li>
-              <Link to="/courseloadtester">Course Load Tester</Link>
-            </li>
-            <li>
-              <Link to="/coffeechat">Coffee Chat</Link>
-            </li>
-          </ul>
-        </nav> */}
         <Routes>
           <Route path="/" element={<LandingPage />} />
           <Route path="/courseSearch" element={<CourseSearch />} />
-          <Route path="/savedCourses" element={<SavedCourses />} />
+          {/* <Route path="/savedCourses" element={<SavedCourses />} /> */}
           <Route path="/courseLoadTester" element={<CourseLoadTester />} />
           <Route path="/coffeeChat" element={<CoffeeChat />}>
           <Route path="/coffeeChat/register" element={<RegisterCoffeeChat />} />
