+++ src/screens/CoffeeChat/contactCard.css
@@ -1,0 +1,112 @@
+.student-detail-container {
+    background-color: #EDECE4;
+    min-height: 100vh;
+    padding: 40px;
+    font-family: 'Itim', Helvetica;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+  
+.page-title {
+    font-family: 'Itim';
+    font-size: 60px;
+    font-weight: bold;
+    color: #66451C;
+    margin-bottom: 30px;
+    text-align: center;
+}
+  
+.student-header {
+    display: flex;
+    justify-content: flex-start;
+    margin-bottom: 10px;
+    width: 90%;
+    max-width: 900px;
+    height: 38px;
+}
+  
+.student-label {
+    background-color: #3B6838;
+    color: #fff;
+    padding: 5px 20px;
+    border-radius: 20px;
+    font-weight: bold;
+    font-family: 'Itim';
+    font-size: 22px;
+}
+  
+.student-card-wrapper {
+    background-color: #3B6838;
+    position: relative;
+    border-radius: 20px;
+    padding: 20px 20px 20px 20px;
+    width: 90%;
+    max-width: 900px;
+    margin: 20px auto;
+    height: 400px;
+}
+  
+.student-content {
+    height: 360px;
+    background-color: #D0E1D1;
+    border-radius: 15px;
+    padding: 10px;
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+    gap: 20px;
+    align-items: center;
+}
+  
+.info-section {
+    padding: 10px;
+    display: flex;
+    flex: 0 0 45%;
+    flex-direction: column;
+    align-items: center;
+    margin: 10px;
+    max-width: none;
+}
+  
+.info-title {
+    background-color: #3B6838;
+    color: #fff;
+    padding: 5px 15px;
+    border-radius: 15px;
+    font-size: 18px;
+    font-family: Helvetica;
+    margin-bottom: 10px;
+    height: 35px;
+}
+  
+.info-box {
+    background-color: #3B6838;
+    color: #fff;
+    padding: 15px;
+    border-radius: 10px;
+    text-align: center;
+    font-size: 18px;
+    width: 250px;
+    width: 100%;
+    height: 150px;
+}
+  
+.back-button {
+    background-color: #E8FABF;
+    color: #000;
+    padding: 10px 25px;
+    border: 1px solid black;
+    border-radius: 25px;
+    font-weight: bold;
+    margin: -100px auto 100px auto;
+    cursor: pointer;
+    font-size: 22px;
+    width: 400px;
+    max-width: 900px;
+    display: block;
+}
+  
+.back-button:hover {
+    background-color: #e0e0b2;
+}
