+++ src/screens/CourseLoadTester/CourseLoadTester.css
@@ -1,0 +1,66 @@
+body {
+    margin: 0;
+    padding: 0;
+    font-family: 'Itim';
+    background-color: #efeee7;
+  }
+  
+  .page-page {
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: row;
+    justify-content: center;
+    padding: 24px;
+    background-color: #efeee7;
+  }
+  
+  .page-name {
+    font-size: 60px;
+    font-weight: bold;
+    margin-bottom: 24px;
+  }
+  
+  .main-container {
+    display: flex;
+    flex-direction: row;
+    padding: 30px; 
+  }
+
+  .course-container {
+    width: 50%;
+    height: 400px;
+    border: 2px solid #3B6838;
+    background-color: #3B6838;
+    position: relative;
+    text-align: center;
+  }
+
+  .tester-container {
+    width: 50%;
+    height: 400px;
+    text-align: center;
+  }
+  
+  .container-label {
+    width: 50%;
+    height: 400px;
+    border: 2px solid #3B6838;
+    background-color: #3B6838;
+    position: relative;
+    text-align: center;
+  }
+
+  .card-card {
+    position: absolute;
+    background-color: #D0E1D1;
+    padding: 8px 16px;
+    border: 2px solid #D0E1D1;
+    border-radius: 25px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    cursor: move;
+    width: 100px;
+    height: 220 px;
+    text-align: center;
+
+  }
