--- src/screens/SavedCourses.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const SavedCourses: React.FC = () => {
-  return <h1>Chat Here</h1>;
-};
-
-export default SavedCourses;
